+++ frontend/src/api/internal/manage/video.ts
@@ -2,7 +2,8 @@ import { apiClient } from 'lib/axios/internal'
 import { cookieHeader } from 'lib/config'
 import { ApiOut, apiOut } from 'lib/error'
 import { Req } from 'types/global'
-import { SearchParms, Video, VideoUpdateIn } from 'types/internal/media'
+import { SearchParms, Video } from 'types/internal/media'
+import { VideoUpdateIn } from 'types/internal/media/manage'
 import { ErrorOut } from 'types/internal/other'
 import { apiManageVideo, apiManageVideos } from 'api/uri'
 import { camelSnake } from 'utils/functions/convertCase'
