+++ frontend/src/components/templates/manage/blog/create.tsx
@@ -1,7 +1,7 @@
 import { useState, ChangeEvent } from 'react'
 import dynamic from 'next/dynamic'
 import { Channel } from 'types/internal/channel'
-import { BlogIn } from 'types/internal/media'
+import { BlogIn } from 'types/internal/media/manage'
 import { Option } from 'types/internal/other'
 import { postBlogCreate } from 'api/internal/manage/create'
 import { FetchError } from 'utils/constants/enum'
