+++ frontend/src/components/templates/manage/music/create.tsx
@@ -1,6 +1,6 @@
 import { useState, ChangeEvent } from 'react'
 import { Channel } from 'types/internal/channel'
-import { MusicIn } from 'types/internal/media'
+import { MusicIn } from 'types/internal/media/manage'
 import { Option } from 'types/internal/other'
 import { postMusicCreate } from 'api/internal/manage/create'
 import { FetchError } from 'utils/constants/enum'
