+++ frontend/src/components/templates/manage/video/edit.tsx
@@ -1,7 +1,8 @@
 import { useState, ChangeEvent } from 'react'
 import { useRouter } from 'next/router'
 import { Channel } from 'types/internal/channel'
-import { Video, VideoUpdateIn } from 'types/internal/media'
+import { Video } from 'types/internal/media'
+import { VideoUpdateIn } from 'types/internal/media/manage'
 import { Option } from 'types/internal/other'
 import { putManageVideo } from 'api/internal/manage/video'
 import { FetchError } from 'utils/constants/enum'
